+++ src/main/java/moriyashiine/aylyth/common/world/generator/trunkplacer/PomegranateTrunkPlacer.java
@@ -1,11 +1,11 @@
-package moriyashiine.aylyth.common.world.generator;
+package moriyashiine.aylyth.common.world.generator.trunkplacer;
 
 import com.google.common.collect.Lists;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import moriyashiine.aylyth.common.registry.ModBlocks;
 import moriyashiine.aylyth.common.registry.ModFeatures;
-import moriyashiine.aylyth.common.world.generator.feature.DirectionalTreeNode;
+import moriyashiine.aylyth.common.world.generator.foliageplacer.DirectionalTreeNode;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.PillarBlock;
 import net.minecraft.util.math.BlockPos;
